servergpu/requirements.txt
Flask
google-genai
SpeechRecognition
